Refactor gulp to handle logos
