+++ gulpfile.js
@@ -1,115 +1,26 @@
+const requireDir   = require('require-dir');
 const gulp         = require( 'gulp' );
-const sourcemaps   = require( 'gulp-sourcemaps' );
 const watch        = require( 'gulp-watch' );
-const postcss      = require( 'gulp-postcss' );
-const autoprefixer = require( 'autoprefixer' );
-const sass         = require( 'gulp-sass' );
-const sassLint     = require( 'gulp-sass-lint' );
-const fileinclude  = require( 'gulp-file-include' );
-const newer        = require( 'gulp-newer' );
-const imagemin     = require( 'gulp-imagemin' );
-const uglify       = require( 'gulp-uglify' );
-const concat       = require( 'gulp-concat' );
-const rename       = require( 'gulp-rename' );
-const gulpCopy     = require( 'gulp-copy' );
-const replace      = require( 'gulp-replace' );
 const del          = require( 'del' );
 
+// For neatness, most tasks have been split up into separate files
+// Just load them all from ./gulp directory.
+requireDir( './gulp' );
+
+// Task to cleanup the dist directory.
 gulp.task( 'clean-dist', ( cb ) => {
 	return del( ['./dist/**/*', '!./dist/.git' ], cb );
 });
 
-// Compile SASS/SCSS to CSS
-gulp.task( 'sass-compile', () => {
-	return gulp.src( './src/styles/*.scss' )
-		.pipe( sourcemaps.init() )
-		.pipe( sass( { outputStyle: 'compressed' } )
-		.on( 'error', sass.logError ) )
-		.pipe( postcss( [ autoprefixer( { browsers: ['last 3 versions'] } ) ] ) )
-		.pipe( sourcemaps.write('.') )
-		.pipe( gulp.dest( './dist/assets/styles' ) );
-});
-
-// Copy SASS to dist for use in other projects.
-gulp.task( 'sass-copy', () => {
-
-	var copy = [
-		'./src/styles/**/*.scss',
-	];
-
-	var ignore = [
-		'!./src/styles/style-guide.scss'
-	];
-
-	return gulp.src( copy.concat( ignore ) )
-		.pipe( gulpCopy( './dist/assets/sass', { prefix: 2 } ) );
-} );
-
-// Watch for changes in JS/CSS.
+// Watch for changes in HTML/JS/CSS.
 gulp.task( 'watch', () => {
 	gulp.watch( 'src/styles/**/*.scss', ['styles'] );
 	gulp.watch( 'src/js/**/*.js', ['js'] );
 	gulp.watch( 'src/html/**/*.html', ['fileinclude'] );
 });
 
-// JavaScript concatination and compression.
-gulp.task( 'js', () => {
-	return gulp.src( './src/js/**/*.js' )
-		.pipe( concat( 'app.js' ) )
-		.pipe( gulp.dest( 'dist/assets/js' ) )
-		.pipe( rename( 'app.min.js' ) )
-		.pipe( uglify() )
-		.pipe( gulp.dest( 'dist/assets/js' ) );
-} );
-
-// HTML file include
-gulp.task( 'fileinclude', () => {
-	return gulp.src( [ './src/html/**/*.html', '!./src/html/components/*', '!./src/html/parts/*' ] )
-		.pipe( fileinclude( {
-			prefix:   '@',
-			basepath: '@file',
-		} ) )
-		// Stip gap inserted between included files.
-		.pipe( replace( '\n\u2028\u2028\n', '\n' ) )
-		.pipe( gulp.dest( './dist/' ) );
-} );
-
-// Minify images.
-gulp.task( 'images', () => {
-	return gulp.src( './src/images/**/*.{jpg,jpeg,png}')
-		.pipe( newer( 'dist/assets/images' ) )
-		.pipe( imagemin( {
-			progressive: true,
-			svgoPlugins: [
-				{ removeViewBox: false },
-				{ cleanupIDs: false }
-			],
-		} ) )
-		.pipe( gulp.dest( 'dist/assets/images' ) );
-} );
-
-// Minify SVG and write to dest.
-gulp.task( 'svg', () => {
-	return gulp.src( './src/images/**/*.svg')
-		.pipe( newer( 'dist/assets/images' ) )
-		.pipe( imagemin( {
-			progressive: true,
-			svgoPlugins: [
-				{ removeViewBox: false },
-				{ cleanupIDs: false }
-			],
-		} ) )
-		.pipe( gulp.dest( './dist/assets/images' ) )
-} );
-
-gulp.task( 'lint-sass', () => {
-	return gulp.src( './src/styles/**/*.s+(a|c)ss')
-		.pipe( sassLint( { configFile: '.sass-lint.yml' } ) )
-		.pipe( sassLint.format() )
-		.pipe( sassLint.failOnError() )
-});
-
-// Tasks
+// Default task. Do everything.
 gulp.task( 'default', [ 'styles', 'js', 'svg', 'images', 'fileinclude', 'lint' ] );
-gulp.task( 'styles',  [ 'sass-compile', 'sass-copy' ] );
-gulp.task( 'lint',    [ 'lint-sass' ] );
+
+// Build task. Should be run before release.
+gulp.task( 'build',   [ 'clean-dist', 'styles', 'js', 'svg', 'images', 'fileinclude' ] );
