+++ package.json
@@ -21,7 +21,8 @@
     "gulp-sass-lint": "^1.2.0",
     "gulp-sourcemaps": "^1.6.0",
     "gulp-uglify": "^1.5.4",
-    "gulp-watch": "^4.3.8"
+    "gulp-watch": "^4.3.8",
+    "require-dir": "^0.3.0"
   },
   "dependencies": {}
 }
